refactor: move db code to module
Refactor existing DB implementation into a module behind a common
interface. There's plans to support multi-DB layers in the future so
using a common abstraction obviates future work. The new abstraction
still leaks SQLite details, but those will disappear when another
backend arrives and provides error use-cases.
